backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0
piper-tts==1.4.1
huggingface-hub==0.24.0
httpx==0.27.0
python-dotenv==1.0.0
numpy==1.26.4
